# platform/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: 1P Clip API
  version: 0.1.0
  description: |
    OpusClip-class AI video repurposing platform. One contract for the web app,
    SDKs, and the MCP server. Auth: workspace API key as `Authorization: Bearer <key>`.
    Every heavy operation is asynchronous — POST /v1/videos returns 202 with a
    projectId to poll.
servers:
  - url: http://localhost:3001
    description: local docker-compose stack
security:
  - bearerAuth: []
paths:
...
  /v1/projects/{id}:
    get:
      summary: Get a project with per-stage pipeline status
      operationId: getProject
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: project }
        "404": { description: not found }
  /v1/projects/{id}/clips:
    get:
      summary: List a project's clips ranked by virality
      operationId: listClips
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  clips:
                    type: array
                    items: { $ref: "#/components/schemas/Clip" }
...
  /v1/clips/{id}/publish:
    post:
      summary: Schedule a clip for a social platform (metadata prefilled when omitted)
      operationId: publishClip
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform: { type: string, enum: [TIKTOK, YOUTUBE, INSTAGRAM, X] }
                title: { type: string }
                description: { type: string }
                hashtags: { type: string }
                scheduledAt: { type: string, format: date-time }
      responses:
        "201": { description: publication created (stays SCHEDULED until a platform OAuth connection is wired) }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: workspace API key or session JWT }
  schemas:
...
    Clip:
      type: object
      properties:
        id: { type: string }
        startSec: { type: number }
        endSec: { type: number }
        title: { type: string }
        viralityScore: { type: number, description: 0-100 }
        viralityBreakdown:
          type: object
          properties:
            hook: { type: number }
            flow: { type: number }
            value: { type: number }
            trend: { type: number }
        viralityReason: { type: string }
        status: { type: string, enum: [SELECTED, PROCESSING, READY, FAILED] }
        renders: { type: array, items: { type: object } }
...